Navigation of pages has been fixed, Lost the action bar, need David to help with that. Workable demo for TA meeting
