+++ app/src/main/java/softeng/eventplanning/MainActivity.java
@@ -44,7 +44,6 @@ public class MainActivity extends AppCompatActivity {
         asyncT.setsomething(create_user_info);
         asyncT.signupActivity(this);
         b.setOnClickListener(new View.OnClickListener() {
-
             @Override
             public void onClick(View v) {
                 Log.d("CALLED", "yup");
@@ -105,7 +104,8 @@ public class MainActivity extends AppCompatActivity {
 
     }
     public void createEventClicked(MenuItem item){
-        setContentView(R.layout.create_event);
+        Intent createEventChange = new Intent(this,createEvent.class);
+        startActivity(createEventChange);
     }
 
     public void logoutClicked(MenuItem item){
@@ -114,26 +114,23 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void setLocation(View v){
-        setContentView(R.layout.create_event_location);
+        Intent eventLocationChange = new Intent(this, createEventLocation.class);
+        startActivity(eventLocationChange);
     }
 
     public void setDate(View v){
-        setContentView(R.layout.create_event_date);
+        Intent eventDateChange = new Intent(this, createEventDate.class);
+        startActivity(eventDateChange);
     }
 
     public void setTime(View v){
-        setContentView(R.layout.create_event_time);
+        Intent eventTimeChange = new Intent(this, createEventTime.class);
+        startActivity(eventTimeChange);
     }
 
     public void back(View v){
-        setContentView(R.layout.create_event);
-
-    }
-
-    public void searchFilter(View view)
-    {
-        Intent intent = new Intent(this, sFilters.class);
-        startActivity(intent);
+        Intent createEventChange = new Intent(this,createEvent.class);
+        startActivity(createEventChange);
     }
 
 
@@ -155,38 +152,9 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void setupTabsView(View v) {
-
-        setContentView(R.layout.activity_main);
         getSupportActionBar().show();
-
-        TabHost tab = (TabHost) findViewById(R.id.mainTabs);
-        tab.setup();
-
-        TabHost.TabSpec spec2 = tab.newTabSpec("Home");
-        spec2.setIndicator("Home");
-        spec2.setContent(R.id.layout2);
-        tab.addTab(spec2);
-
-        TabHost.TabSpec spec1 = tab.newTabSpec("Current Event");
-        spec1.setIndicator("Current Event");
-        spec1.setContent(R.id.layout1);
-        tab.addTab(spec1);
-
-        TabHost.TabSpec spec3 = tab.newTabSpec("Search");
-        spec3.setIndicator("Search");
-        spec3.setContent(R.id.layout3);
-        tab.addTab(spec3);
-
-        ArrayAdapter adapter = new ArrayAdapter<String>(this, R.layout.searchresultslist, mobileArray);
-        ListView listView = (ListView) findViewById(R.id.searchresults);
-        listView.setAdapter(adapter);
-
-        exp_list = (ExpandableListView) findViewById(R.id.exp_list);
-        Friends = ListData.getInfo();
-        Friend_list = new ArrayList<String>(Friends.keySet());
-
-        adapter2 = new friendadapter(this, Friends, Friend_list);
-        exp_list.setAdapter(adapter2);
+        Intent tabViewChange = new Intent(this, tabView.class);
+        startActivity(tabViewChange);
 
     }
