+++ app/src/main/java/softeng/eventplanning/createEvent.java
@@ -1,0 +1,19 @@
+package softeng.eventplanning;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+/**
+ * Created by jpcerone on 10/3/16.
+ */
+
+public class createEvent extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.create_event);
+
+    }
+
+}
