+++ app/src/main/java/softeng/eventplanning/sFilters.java
@@ -17,7 +17,7 @@ public class sFilters extends AppCompatActivity {
     }
     public void filterBack(View view)
     {
-        Intent intent = new Intent(this, MainActivity.class);
+        Intent intent = new Intent(this, tabView.class);
         startActivity(intent);
     }
